split issue 92 unit test into its own file
